# docs/manual-payments.openapi.yml
openapi: 3.0.3
info:
  title: Anytimebot — Pagos manuales y ventas no agendadas
  version: 1.0.0
  description: |
    Módulo de ventas y cobros manuales. Todos los importes van en **céntimos
    enteros** (`amountCents`) y la moneda es la del tenant en minúsculas.
    Los pagos nunca se borran: se anulan (con motivo) o se reembolsan.
    Stripe solo interviene en pagos online (`CARD_ONLINE` con
    `stripePaymentIntent` real); los métodos manuales jamás llaman a Stripe.
    Autenticación: cookie de sesión NextAuth (dashboard).
servers:
  - url: https://anytimebot.app
    description: Producción
  - url: http://localhost:3000
    description: Desarrollo
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
    OrderItemInput:
      type: object
      required: [description, unitCents]
      properties:
        description: { type: string }
        quantity: { type: integer, minimum: 1, default: 1 }
        unitCents: { type: integer, minimum: 0 }
        taxRateBps: { type: integer, nullable: true, description: "Informativo. 2100 = 21 %." }
    Balance:
      type: object
      properties:
        total: { type: integer }
        paid: { type: integer }
        refunded: { type: integer }
        net: { type: integer }
        due: { type: integer }
        state: { type: string, enum: [UNPAID, PARTIAL, PAID, OVERPAID] }
    Order:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [ISSUED, VOID] }
        currency: { type: string }
        bookingId: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItemInput' }
        payments:
          type: array
          items: { $ref: '#/components/schemas/Payment' }
        receiptNumber: { type: string, nullable: true, example: "REC-2026-000123" }
        balance: { $ref: '#/components/schemas/Balance' }
    Payment:
      type: object
      properties:
        id: { type: string }
        method: { type: string, enum: [CASH, CARD_ONSITE, TRANSFER, BIZUM, OTHER, CARD_ONLINE] }
        amountCents: { type: integer, minimum: 1 }
        currency: { type: string }
        status: { type: string, enum: [COMPLETED, ANNULLED] }
        reference: { type: string, nullable: true }
        refundCents: { type: integer }
        annulledReason: { type: string, nullable: true }
        createdByName: { type: string }
        createdAt: { type: string, format: date-time }
  securitySchemes:
    nextAuthCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
security:
  - nextAuthCookie: []
paths:
  /api/orders:
    get:
      tags: [orders]
      summary: Listar ventas (paginado)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [ISSUED, VOID] } }
      responses:
        '200':
          description: Lista con saldo por orden
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
                  meta:
                    type: object
                    properties:
                      page: { type: integer }
                      pageSize: { type: integer }
                      total: { type: integer }
    post:
      tags: [orders]
      summary: Venta rápida (sin cita)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/OrderItemInput' }
                customerId: { type: string, nullable: true }
                locationId: { type: string, nullable: true }
                note: { type: string, nullable: true }
                payment:
                  type: object
                  nullable: true
                  properties:
                    method: { type: string, enum: [CASH, CARD_ONSITE, TRANSFER, BIZUM, OTHER] }
                    amountCents: { type: integer, minimum: 1 }
                    reference: { type: string, nullable: true }
                    idempotencyKey: { type: string, nullable: true }
      responses:
        '201': { description: Venta creada (con recibo secuencial) }
        '400': { description: Ítems o importes no válidos, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  /api/bookings/{id}/payments:
    get:
      tags: [payments]
      summary: Orden e historial de pagos de una cita
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Cita + orden + balance }
    post:
      tags: [payments]
      summary: Registrar cobro manual en una cita (multi-pago / parcial)
      description: >
        Crea (o reutiliza) la orden de la cita —absorbiendo un cobro Stripe
        histórico como espejo CARD_ONLINE— y añade el pago. Con
        `complete: true` y saldo completo, finaliza la cita (factura + avisos).
        Sincroniza las columnas planas legacy cuando el saldo llega a PAID.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method, amountCents]
              properties:
                method: { type: string, enum: [CASH, CARD_ONSITE, TRANSFER, BIZUM, OTHER] }
                amountCents: { type: integer, minimum: 1 }
                reference: { type: string, nullable: true }
                idempotencyKey: { type: string, nullable: true }
                complete: { type: boolean, default: false }
      responses:
        '200': { description: Pago registrado (data.balance = saldo de la orden) }
        '400': { description: Método o importe no válido }
    delete:
      tags: [payments]
      summary: Anular los cobros manuales de la cita
      description: Nunca toca pagos CARD_ONLINE (esos se reembolsan en Stripe).
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Cobros anulados }
        '409': { description: Solo hay pagos de Stripe }
  /api/payments/{id}:
    get:
      tags: [payments]
      summary: Pago individual (vista de recibo)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Pago con su orden }
    post:
      tags: [payments]
      summary: Anular o reembolsar un pago
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, reason]
              properties:
                action: { type: string, enum: [annul, refund] }
                reason: { type: string }
                amountCents:
                  type: integer
                  description: Solo para refund; por defecto reembolsa todo lo pendiente.
      responses:
        '200': { description: Pago actualizado (nunca borrado) }
        '400': { description: Reembolso mayor que lo pendiente }
        '409': { description: Pago ya anulado o con reembolsos previos (annul) }
...